refactor(web): retire legacy Google module alias shim
Rewrite ~30 production and test imports from dead @web/auth/google/hooks,
state, and util specifiers to their real @web/auth/providers paths. Update
mock.module keys to match. Remove legacy-google-aliases.ts and its tsconfig,
dev, build, and test-runner registrations.
