+++ packages/web/src/components/AuthModal/AuthModal.test.tsx
@@ -23,7 +23,7 @@ import { registerUseStartGoogleAuthorizationForTests } from "@web/auth/google/au
 import {
   resetGoogleAvailabilityForTests,
   setGoogleAvailabilityForTests,
-} from "@web/auth/google/hooks/useIsGoogleAvailable/useIsGoogleAvailable";
+} from "@web/auth/providers/useIsProviderAvailable";
 import { AuthModal } from "./AuthModal";
 import { AuthModalProvider } from "./AuthModalProvider";
 import { useAuthModal, validateAuthSearch } from "./hooks/useAuthModal";
