+++ packages/web/src/components/Sidebar/SidebarStatusBar.tsx
@@ -1,10 +1,10 @@
 import { type FC } from "react";
-import { useConnectGoogle } from "@web/auth/google/hooks/useConnectGoogle/useConnectGoogle";
 import {
   getSidebarSyncStatus,
   SSE_DEGRADED_STATUS,
-} from "@web/auth/google/hooks/useConnectGoogle/useConnectGoogle.util";
-import { useGoogleSyncRefreshSnapshot } from "@web/auth/google/state/google.sync.refresh";
+} from "@web/auth/providers/connect.util";
+import { useGoogleSyncRefreshSnapshot } from "@web/auth/providers/sync.refresh";
+import { useConnectGoogle } from "@web/auth/providers/useConnectProvider";
 import { useShortcutWriteLocked } from "@web/billing/useBillingWriteLock";
 import { SYNC_STATUS_VARIANT_CLASSNAME } from "@web/calendars/sync-status.types";
 import { useHasPendingEventMutations } from "@web/events/mutations/useEventPending";
