+++ packages/web/src/views/Week/components/Grid/Grid.tsx
@@ -2,11 +2,11 @@ import { type FC, useMemo } from "react";
 import { YEAR_MONTH_DAY_FORMAT } from "@core/constants/date.constants";
 import { type Dayjs } from "@core/util/date/dayjs";
 import { shouldShowContextualLoadError } from "@web/api/util/api.util";
-import { useConnectGoogle } from "@web/auth/google/hooks/useConnectGoogle/useConnectGoogle";
 import {
   isFirstImportFailed,
   isFirstImportInProgress,
-} from "@web/auth/google/hooks/useConnectGoogle/useConnectGoogle.util";
+} from "@web/auth/providers/connect.util";
+import { useConnectGoogle } from "@web/auth/providers/useConnectProvider";
 import { useWeekEventViewModel } from "@web/events/queries/useWeekEventsQuery";
 import { selectGridDraft, useDraftStore } from "@web/events/stores/draft.store";
 import { EventGrid, isEventGridLoading } from "@web/grid/components/EventGrid";
